feat(docs): write docs, publish initial major
adds docs, CI/CD

BREAKING CHANGE: publish initial 1.0.0
