+++ bench/index.js
@@ -1,28 +1,33 @@
 const { Suite } = require('benchmark')
 const React = require('react')
 const { renderToStaticMarkup } = require('react-dom/server')
+const hyperscript = require('hyperscript')
 const hyposcript = require('../')
 
-function bench(name) {
-	console.log(`\n# ${name}`);
-	const suite = new Suite();
-	const previous = suite.add.bind(suite);
-	suite.on('cycle', e => console.log('  ' + e.target));
-	suite.add = (name, runner) => previous(name.padEnd(16), runner);
-	return suite;
+function bench (name) {
+  console.log(`\n# ${name}`)
+  const suite = new Suite()
+  const previous = suite.add.bind(suite)
+  suite.on('cycle', e => console.log('  ' + e.target))
+  suite.add = (name, runner) => previous(name.padEnd(16), runner)
+  return suite
 }
 
-function createApp(h) {
-  return h('div', { className: 'foo' },
-    h('ul', {},
+function createApp (h) {
+  return h(
+    'div',
+    { className: 'foo' },
+    h(
+      'ul',
+      {},
       h('li', {}, 'a'),
-      h('li', {}, 'a'),
-      h('li', {}, 'a'),
-      h('li', {}, 'a'),
-      h('li', {}, 'a'),
-      h('li', {}, 'a'),
-      h('li', {}, 'a'),
-      h('li', {}, 'a'),
+      h('li', {}, 'b'),
+      h('li', {}, 'c'),
+      h('li', {}, 'd'),
+      h('li', {}, 'e'),
+      h('li', {}, 'f'),
+      h('li', {}, 'g'),
+      h('li', {}, 'h')
     )
   )
 }
@@ -31,7 +36,10 @@ bench('render')
   .add('react', () => {
     renderToStaticMarkup(createApp(React.createElement))
   })
+  .add('hyperscript', () => {
+    createApp(hyperscript).outerHTML
+  })
   .add('hyposcript', () => {
-    renderToStaticMarkup(createApp(hyposcript.h))
+    createApp(hyposcript.h)
   })
   .run()
