+++ bench/package.json
@@ -2,6 +2,7 @@
   "private": true,
   "devDependencies": {
     "benchmark": "^2.1.4",
+    "hyperscript": "^2.0.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   }
